==> Notebooks/2.-analisis_sentiemientos.ipynb <==
# Análisis de Sentimientos

Vamos a analizar los textos de las reseñas de `user_reviews` para crear la columna `sentiment_analysis`. Esta columna tendrá un valor de 0 si es negativo, 1 si es neutral, o 2 si es positivo. En caso de que no haya una reseña escrita, se le asignará un valor de 1.


### Importación de las librerias

import pandas as pd
from nltk.sentiment import SentimentIntensityAnalyzer
import nltk
nltk.download('vader_lexicon')

### Lectura del CSV

# leemos el archivo
user_reviews = pd.read_csv("..\\Datsets\\CSV\\user_reviews.csv")

### Función para Asignar Etiquetas

Esta función se empleará para asignar etiquetas numéricas según el umbral del analizador de sentimientos SentimentIntensityAnalyzer.

Esta función recibe un puntaje de sentimiento como entrada y asigna etiquetas numéricas según umbrales predefinidos.

**Parámetros:**
- `score` (float): Puntaje de sentimiento obtenido del análisis.

**Retorno:**
- `int`: Etiqueta numérica asignada, donde:
- `2` indica sentimiento positivo,
- `0` indica sentimiento negativo, y
- `1` indica neutralidad.


def asignar_etiqueta(sentimiento):
    """
    Función para Asignar Etiquetas Numéricas a los Sentimientos

    Esta función recibe un puntaje de sentimiento como entrada y asigna etiquetas numéricas según umbrales predefinidos.

    Parámetros:
    - sentimiento (float): Puntaje de sentimiento obtenido del análisis.

    Retorno:
    - int: Etiqueta numérica asignada, donde:
           - 2 indica sentimiento positivo,
           - 0 indica sentimiento negativo, y
           - 1 indica neutralidad.

    Ejemplo de Uso:
    ```python
    puntaje_sentimiento = analizar_sentimiento(texto)
    etiqueta_asignada = asignar_etiqueta(puntaje_sentimiento)
    print(f"Etiqueta Asignada: {etiqueta_asignada}")
    ```
    """
    if sentimiento >= 0.05:
        return 2  # Sentimiento positivo
    elif sentimiento <= -0.04:
        return 0  # Sentimiento negativo
    else:
        return 1  # Neutral


Las filas de la columna 'review' que contengan valores nulos se van a reemplazar por cadenas vacías
para que el analizador de sentimientos les dé un valor de 0.

# Rellenar valores nulos con una cadena vacía
user_reviews['review'].fillna('', inplace=True)

### Modelo

Utilizaremos la librería NLTK (Natural Language Toolkit) para realizar un análisis de sentimientos de las revisiones de usuarios.

1. **Inicializar el analizador de sentimientos de NLTK:**
   Aquí, se crea una instancia del analizador de sentimientos SentimentIntensityAnalyzer de NLTK. Este analizador asigna un puntaje de sentimiento a cada revisión.

2. **Aplicar el análisis de sentimiento y asignar etiquetas numéricas:**
   Se aplica el análisis de sentimiento a la columna 'review' del DataFrame 'user_reviews'. El puntaje de sentimiento se obtiene mediante la función `sid.polarity_scores(x)['compound']`, donde 'x' es cada revisión. Luego, se utiliza la función `asignar_etiqueta` para asignar etiquetas numéricas (positivo, negativo, neutral) basadas en los puntajes de sentimiento.

3. **Borrar la columna 'review':**
   Finalmente, se elimina la columna original 'review' del DataFrame 'user_reviews', ya que ahora tenemos la información de sentimiento y ya no necesitamos las revisiones individuales.

# Inicializar el analizador de sentimientos de NLTK
sid = SentimentIntensityAnalyzer()

# Aplicar el análisis de sentimiento y asignar etiquetas numéricas
user_reviews['sentiment_score'] = user_reviews['review'].apply(lambda x: sid.polarity_scores(x)['compound'])
user_reviews['sentiment_analysis'] = user_reviews['sentiment_score'].apply(asignar_etiqueta)

# Borramos la columa reviews
user_reviews.drop(columns=["review"],inplace=True)


### Exportar a CSV


Exportar a CSV y reemplazar el archivo 'user_reviews' con este, que ahora incluye dos columnas adicionales relacionadas con el análisis de sentimientos.

# Convertir el DataFrame 'user_reviews' a CSV
user_reviews.to_csv("..\\Datsets\\CSV\\user_reviews.csv", index=False)
