==> Notebooks/3.-play_time_genre.ipynb <==
# PlayTimeGenre

Crear una función que tenga como entrada una cadena de texto representando un género y nos devuelva el año con más horas jugadas para dicho género. Para hacer esto se va a crear el CSV con la información necesaria para realizar la consulta, ya que se debe deployar en Render y los datasets son muy pesados para cargarlos.

def PlayTimeGenre( genero : str ): Debe devolver año con mas horas jugadas para dicho género.
Ejemplo de retorno: {"Año de lanzamiento con más horas jugadas para Género X" : 2013}


### Importamos las librerías

import pandas as pd
import ast

### Lectura del archivo CSV

steam_games = pd.read_csv("..\\Datsets\\CSV\\steam_games.csv")
user_items =pd.read_csv("..\\Datsets\\CSV\\user_items.csv")

### Creación del CSV para la función PlayTimeGenre

La información que necesitamos incluye los géneros, el año de lanzamiento y el tiempo jugado en total por cada juego. Para ello, vamos a utilizar las columnas 'id', 'genres', 'release_year' del conjunto de datos `steam_games`, y las columnas 'item_id' y 'playtime_forever' del conjunto de datos `user_items`. Vamos a concatenar los DataFrames a través de las columnas 'id' e 'item_id'.

1. **Columnas de steam_games:**
    - 'id'
    - 'genres'
    - 'release_year'

2. **Columnas de user_items:**
    - 'item_id'
    - 'playtime_forever'

Vamos a realizar la concatenación utilizando las columnas 'id' e 'item_id'.


# Combinar DataFrames utilizando las variables en español
tiempo_juego_genero = pd.merge(steam_games[['id', 'genres', 'release_year']],
                               user_items[['item_id', 'playtime_forever']],
                               left_on='id', right_on='item_id', how='inner')

# Eliminar las columnas duplicadas 'item_id' e 'id'
tiempo_juego_genero = tiempo_juego_genero.drop(columns=['item_id'], axis=1)


tiempo_juego_genero.head()

Agrupar la columna 'playtime_forever' por cada 'id' para obtener el tiempo total jugado por cada juego

#Agrupamos por playtime_forever por id
tiempo_juego_genero = tiempo_juego_genero.groupby(['genres', 'release_year', 'id'])['playtime_forever'].sum().reset_index()

tiempo_juego_genero.head()

Expandimos las listas en la columna 'genres'

#Convertimos la columna genres a tipo lista
tiempo_juego_genero['genres'] = tiempo_juego_genero['genres'].apply(ast.literal_eval)
#Expandimos las listas en columnas
tiempo_juego_genero = tiempo_juego_genero.explode('genres').reset_index(drop=True)

tiempo_juego_genero.head()

Agrupamos por 'genres' y 'release_year', sumando las horas jugadas para cada combinación.

#Agrupamos por genero y año de lanzamiento
tiempo_juego_genero = tiempo_juego_genero.groupby(['release_year', 'genres'])['playtime_forever'].sum().reset_index()

tiempo_juego_genero.head()

El resultado es un DataFrame que contiene la información de cuántas horas se han jugado por género y año de lanzamiento. Convertiremos este DataFrame a un archivo CSV para alimentar la API.

# Guardar el DataFrame en un archivo CSV
tiempo_juego_genero.to_csv('..\\Datsets\\Archivos_API\\PlayTimeGenre.csv', index=False)


Creación de la función para consultar el año con más horas jugadas por género

def PlayTimeGenre(genero):
    """
    Función que toma como entrada un género y devuelve el año con la máxima cantidad de horas jugadas para ese género.

    Parámetros:
    - genero (str): Género para el cual se busca el año con más horas jugadas.

    Ejemplo de retorno:
    - Para un género existente: 2013
    - Para un género no encontrado: "No se encontraron datos para el género 'Género X'"
    """
    # Filtra el DataFrame para el género especificado

    df_genero = tiempo_juego_genero[tiempo_juego_genero['genres'] == genero]

    # Verifica si el género está presente en el DataFrame
    if not df_genero.empty:
        # Encuentra el año con la máxima cantidad de horas jugadas
        anio_mas_jugado = df_genero.loc[df_genero['playtime_forever'].idxmax()]['release_year']
        return f'Año de lanzamiento con más horas jugadas para el género {genero} : {int(anio_mas_jugado)}'
    else:
        return f"No se encontraron datos para el género '{genero}'"
