==> Notebooks/4.-user_for_genre.ipynb <==
# UserForGenre

Definir una función que identifique el usuario con la mayor cantidad de horas jugadas en el género especificado y proporcione una lista que muestre la acumulación de horas jugadas por año.

def UserForGenre( genero : str ): 
Ejemplo de retorno: {"Usuario con más horas jugadas para Género X" : us213ndjss09sdf, "Horas jugadas":[{Año: 2013, Horas: 203}, {Año: 2012, Horas: 100}, {Año: 2011, Horas: 23}]}

### Importamos las librerías

import pandas as pd
import ast

### Lectura de los archivos CSV

steam_games = pd.read_csv("..\\Datsets\\CSV\\steam_games.csv")
user_items =pd.read_csv("..\\Datsets\\CSV\\user_items.csv")

La información necesaria para esta consulta se encuentra en los dataframes `steam_games` y `user_items`. Del dataframe `steam_games`, requerimos las columnas 'genres' y 'release_year', mientras que del dataframe `user_items` necesitamos los datos de 'playtime_forever' y 'user_id'. Para lograr esto, procederemos a concatenar los dataframes utilizando la columna 'item_id'.

# Unir los DataFrames usando las columnas 'id' en steam_games y 'item_id' en user_items
df_UserForGenre = pd.merge(user_items, steam_games, left_on='item_id', right_on='id', how='inner')

# Seleccionar las columnas requeridas
df_UserForGenre = df_UserForGenre[['genres', 'release_year', 'playtime_forever', 'user_id']]

df_UserForGenre.head()

Expandimos las listas de la columna genres en filas.

# Verificar si 'genres' contiene NaN antes de aplicar ast.literal_eval
df_UserForGenre['genres'] = df_UserForGenre['genres'].apply(lambda x: ast.literal_eval(x) if pd.notna(x) else [])
#Expandimos las listas en columnas
df_UserForGenre = df_UserForGenre.explode('genres').reset_index(drop=True)
df_UserForGenre.head()

Buscamos al usuario que tiene la mayor cantidad de horas jugadas para cada género

# Encontrar el user_id con la mayor cantidad de horas jugadas por género
jugador_mas_horas = df_UserForGenre.groupby(['genres', 'user_id'])['playtime_forever'].sum().reset_index()

# Encontrar el índice del user_id con la mayor cantidad de horas jugadas por género
indice = jugador_mas_horas.groupby('genres')['playtime_forever'].idxmax()

# Filtrar el DataFrame original usando los índices encontrados
jugador_mas_horas = jugador_mas_horas.loc[indice]

jugador_mas_horas.head()

Concatenamos los DataFrames jugador_mas_horas y df_UserForGenre para retener únicamente al jugador que ha acumulado la mayor cantidad de horas por género del DataFrame df_UserForGenre. Además, conservamos la información sobre la cantidad de horas jugadas por año de lanzamiento.

# Concatenamos los DataFrames
df_UserForGenre = pd.merge(df_UserForGenre, jugador_mas_horas[['genres', 'user_id']], on=['genres','user_id'])
df_UserForGenre.head()

Filtramos los valores para eliminar las entradas donde la cantidad de horas jugadas (playtime_forever) es igual a 0.

df_UserForGenre = df_UserForGenre[df_UserForGenre['playtime_forever']>0]
df_UserForGenre.head()

Agrupamos por release_year

df_UserForGenre = df_UserForGenre.groupby(['genres', 'release_year','user_id'])['playtime_forever'].sum().reset_index()
df_UserForGenre.head()

### Convertimos a CSV

#Convertimos a CSV
df_UserForGenre.to_csv('..\\Datsets\\Archivos_API\\UserForGenre.csv', index=False)

### Función

def UserForGenre(genre):
    # Filtrar el DataFrame por el género dado
    genero = df_UserForGenre[df_UserForGenre['genres'] == genre]

    # Crear una lista de diccionarios para las horas jugadas por año
    horas_anio = [{'Año': int(year), 'Horas': int(playtime)} for year, playtime in genero[['release_year', 'playtime_forever']].values]

    # Crear el diccionario de salida
    resultado = {
        'Usuario con más horas jugadas para género '+ genre: genero.iloc[0]['user_id'],
        "Horas jugadas": horas_anio
    }

    return resultado


genre_result = UserForGenre('Action')
print(genre_result)
