==> Notebooks/5.-users_recommend.ipynb <==
# UserRecommend

La función UsersRecommend tiene como objetivo proporcionar a los usuarios la información sobre los tres juegos más recomendados para un año específico. La función toma como entrada un año y devuelve el top 3 de juegos más recomendados por la comunidad de usuarios. Solo se considerarán comentarios positivos de las reseñas. 

Ejemplo de retorno: [{"Puesto 1" : X}, {"Puesto 2" : Y},{"Puesto 3" : Z}]

### Importación de librerias

import pandas as pd

### Lectura de los CSV

steam_games = pd.read_csv("..\\Datsets\\CSV\\steam_games.csv")
user_reviews =pd.read_csv("..\\Datsets\\CSV\\user_reviews.csv")

### Creación de DataFrame

Para la implementación de esta función, se utilizarán dos DataFrames: user_reviews y steam_games. Estos DataFrames se concatenarán  para obtener las columnas necesarias que serán utilizadas en la función.

# Combinar DataFrames basado en las columnas id y item_id
df_UsersRecommend = pd.merge(user_reviews, steam_games, left_on='item_id', right_on='id', how='inner')
df_UsersRecommend.head()

Seleccionamos las columnas que van a ser necesarias 

# Seleccionar las columnas necesarias
selected_columns = ['app_name', 'posted_year', 'recommend', 'sentiment_analysis']
df_UsersRecommend = df_UsersRecommend[selected_columns]
df_UsersRecommend.head()

Filtrar los comentarios positivos y neutrales

#Filtramos los comentarios positivos y neutrales
df_UsersRecommend =df_UsersRecommend[df_UsersRecommend['sentiment_analysis']!=0]
df_UsersRecommend.head()

Agrupar por el nombre del juego y la fecha de la recomendación

#Agrupamos por app_name , posted_year y sumamos la cantidad de recomendaciones
df_UsersRecommend = df_UsersRecommend.groupby(['app_name', 'posted_year']).agg({'recommend': 'sum'}).reset_index()
df_UsersRecommend = df_UsersRecommend.sort_values('recommend',ascending=False)
df_UsersRecommend.head()

### Convertir el DataFrame a CSV

df_UsersRecommend.to_csv('..\\Datsets\\Archivos_API\\UsersRecommend.csv', index=False)


## Función UserRecommend

def UsersRecommend(anio):
    """
    Devuelve el top 3 de juegos MÁS recomendados por usuarios para el año dado.
    
    Args:
    - anio (int): El año para el cual se desean obtener las recomendaciones.

    Returns:
    - dict: Un diccionario que contiene el top 3 de juegos recomendados en el formato:
        {"Puesto 1": "Nombre del Juego1", "Puesto 2": "Nombre del Juego2", "Puesto 3": "Nombre del Juego3"}
        En caso de no haber recomendaciones para el año especificado, devuelve un mensaje indicando esto.
    """

    # Filtrar las recomendaciones para el año dado y recomendaciones positivas/neutrales
    recomendaciones = df_UsersRecommend[df_UsersRecommend['posted_year'] == anio]

    # Verificar si hay revisiones para el año dado
    if recomendaciones.empty:
        return f"No hay recomendaciones para el año {anio}"

    # Ordenar en orden descendente por la cantidad de recomendaciones
    recomendaciones = recomendaciones.sort_values('recommend', ascending=False)

    # Crear una única línea de resultado
    resultado = {
        "Puesto 1": recomendaciones.iloc[0]['app_name'],
        "Puesto 2": recomendaciones.iloc[1]['app_name'],
        "Puesto 3": recomendaciones.iloc[2]['app_name']
    }

    return resultado


UsersRecommend(2011)
