==> Notebooks/6.-users_worst_developer.ipynb <==
# UsersWorstDeveloper

Se busca la creación de una función que proporcione el top 3 de desarrolladoras con juegos menos recomendados por usuarios para un año específico. La función toma como entrada el año y devuelve las 3 desarrolladoras con los juegos menos recomendados, considerando solo aquellos con recomendaciones "False" y comentarios negativos.

Ejemplo de retorno : [{"Puesto 1" : X}, {"Puesto 2" : Y},{"Puesto 3" : Z}]


### Importación de librerías

import pandas as pd

### Lectura de los archivos CSV

steam_games = pd.read_csv("..\\Datsets\\CSV\\steam_games.csv")
user_reviews =pd.read_csv("..\\Datsets\\CSV\\user_reviews.csv")

### Creacion del DataFrame

Para la implementación de la función `UsersWorstDeveloper`, se utilizarán dos archivos: "steam_games" y "user_reviews". Estos archivos contienen información para la función, las cuales son : desarrollador (`developer`), recomendación (`recommend`), análisis de sentimiento (`sentiment_analysis`), y el año de publicación (`posted_year`).

# Combinar dataframes en base a las columnas id e item_id
df_UserWorstDeveloper = pd.merge(steam_games, user_reviews, left_on='id', right_on='item_id')

# Seleccionar columnas 
df_UserWorstDeveloper = df_UserWorstDeveloper[['developer', 'recommend', 'sentiment_analysis','posted_year']]

Seleccionar únicamente aquellos juegos que cuenten con comentarios negativos del campo sentiment_analysis

#Filtramos los comentarios negativos
df_UserWorstDeveloper = df_UserWorstDeveloper[df_UserWorstDeveloper['sentiment_analysis'] == 0]

Agrupar el número de recomendaciones por desarrollador y por año

#Agrupamos las recomendaciones por desarrollador y por año
df_UserWorstDeveloper = df_UserWorstDeveloper.groupby(['developer', 'posted_year'])['recommend'].sum().reset_index()
df_UserWorstDeveloper.head()

### Convertimos el DataFrame a CSV

df_UserWorstDeveloper.to_csv('..\\Datsets\\Archivos_API\\UsersWorstDeveloper.csv', index=False)

### Función UserWorstDeveloper

def UserWorstDeveloper(anio):
    desarrollador = df_UserWorstDeveloper[df_UserWorstDeveloper['posted_year']== anio]
    if not desarrollador.empty:
        # Encuentra los años con menos recomendaciones
        desarrollador = desarrollador.nsmallest(3, 'recommend')
        resultado = [{"Puesto 1" : desarrollador.iloc[0]['developer']}, 
                  {"Puesto 2" : desarrollador.iloc[1]['developer']},
                  {"Puesto 3" : desarrollador.iloc[2]['developer']}]
        return resultado
    else:
        return f"No se encontraron datos para el año '{anio}'"
    

UserWorstDeveloper(2012)
